Land the #2951 work left in the shared tree: the addition-manifold benches, the 4-layer target port, the grokking movie and the complex128 charge bases
- bench/mpd_llm_addition_{harvest,manifolds,interventions}_2951.py: the three phases
  of the manifold-native two-digit addition analysis (harvest a pretrained LM's
  residual stream; fit each problem variable's curve by gamfit REML; move a
  variable along its fitted curve and read the answer).
- bench/mpd_vpd4l_2951.py: VPD's 4-layer LlamaSimpleMLP target ported from
  param_decomp's declared forward and checked against the run's logged validation
  loss, with the gate-manifold decomposition measured against it.
- bench/mpd_llm_digit_{charges,layer_charges}_2951.py: build the charge basis in
  complex128; the untyped tensor defaulted to complex64.
- experiments/issue-2951/movies/modadd_grokking.{mp4,gif}: the grokking movie
  (modadd_disc.mp4 was already here).

Not landed: bench/mpd_vpd4l_identifiability_2951.py measures trained frames of the
orthogonal-frame method 2ab638a132 deleted and imports its deleted benches.

+++ bench/mpd_llm_digit_charges_2951.py
@@ -144,7 +144,7 @@ def main():
                 F = harmonics(lp)
                 gains.append((F * F0.conj()).sum(0) / F0.abs().pow(2).sum(0))
             g = torch.stack(gains).cpu()  # M x 5
-            basis = torch.tensor([[complex(math.cos(n * t), math.sin(n * t)) for t in thetas] for n in charges])
+            basis = torch.tensor([[complex(math.cos(n * t), math.sin(n * t)) for t in thetas] for n in charges], dtype=torch.complex128)
             spec = basis @ g / M  # charges x 5
             entry = {}
             for jj in range(5):
